Moves DB uniqueness initialization to DBController, modernize
